// types/employee.ts
export type Employee = {
  id: string
  fullName: string
  designation: string
  department: string
  mobile: string
  email: string
  extension: string
}
export type AdminEmployee = {
  id: number
  first_name: string
  last_name: string
  designation: string
  department: string
  cell_phone: string
  email: string
  work_phone: string
  ip_address: string
  created_by: number
  created_date: string
  modified_by: number
  modified_date: string
}